feat: Inicializar la estructura del proyecto backend con módulos de autenticación, gestión de usuarios y citas.

+++ src/main/java/com/neodent/cita/model/HorarioOdontologo.java
@@ -31,7 +31,7 @@ public class HorarioOdontologo {
     private Sede sede;
 
     @Column(name = "dia_semana", nullable = false)
-    private Integer diaSemana;
+    private Byte diaSemana;
 
     @Column(name = "hora_inicio", nullable = false)
     private LocalTime horaInicio;
